requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
dotenv==0.9.9
exceptiongroup==1.2.2
executing==2.2.0
fonttools==4.56.0
greenlet==3.1.1
idna==3.10
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==4.9.4
matplotlib==3.10.0
matplotlib-inline==0.1.7
matplotsoccer==0.0.8
mplsoccer==1.4.0
multimethod==1.12
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pandera==0.17.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.3.0
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
socceraction==1.5.3
SQLAlchemy==2.0.40
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
typeguard==4.4.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
wrapt==1.17.2